requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
grpcio==1.36.1
grpcio-tools==1.36.1
idna==2.10
jsonschema==3.2.0
lxml==4.6.2
nest-asyncio==1.5.1
protobuf==3.15.5
requests==2.25.1
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
